refactor: shared hooks cartCount and cartItems
