+++ src/components/CartProduct/CartProduct.module.css
@@ -8,6 +8,7 @@
   box-shadow: 18px 18px 36px 2px rgba(211, 209, 216, 0.65);
   padding: 40px 60px;
   border-radius: 19px;
+  width: 1000px;
 }
 
 .light {
