+++ src/pages/ProductPage/ProductPage.module.css
@@ -51,6 +51,13 @@
   flex-direction: row;
   justify-content: space-between;
   align-items: center;
+  height: 100px;
+}
+
+.actions {
+  display: flex;
+  justify-content: flex-start;
+  align-self: center;
 }
 
 .price {
